Fix input read when closing bubbletea app
When we close the snippets screen, the last input keystroke is read by
the bubbletea framework, so we lose it and users have to press the key
again.

To fix this problem, we keep the last input keystroke in a buffer while
bubblea app is running (detached mode).
When the snippets screen is closed, we attach our app back to the stdin
and send the buffered input to the stdin channel again.